Create Dictionaries_Part_2
Dictionary Manipulation (1)
If you know how to access a dictionary, you can also assign a new value to it. 
To add a new key-value pair to europe you can use something like this:

europe['iceland'] = 'reykjavik'

-----------------------------------

Dictionary Manipulation (2)
Somebody thought it would be funny to mess with your accurately generated dictionary. An adapted version of the europe dictionary is available in the script on the right.

Can you clean up? Do not do this by adapting the definition of europe, but by adding Python commands to the script to update and remove key:value pairs.

------------------------------

Dictionariception
Remember lists? They could contain anything, even other lists. Well, for dictionaries the same holds. Dictionaries can contain key:value pairs where the values are again dictionaries.

As an example, have a look at the script where another version of europe - the dictionary you've been working with all along - is coded. The keys are still the country names, but the values are dictionaries that contain more information than just the capital.

It's perfectly possible to chain square brackets to select elements. To fetch the population for Spain from europe, for example, you need:
europe['spain']['population']

------------------------------